move format-tree to utils
