+++ addon/addon/utils/format-tree.ts
